Add files via upload
It is my first project on data Analytics. In this project I worked on the dataset that I get from Kaggle.
I analyze the data by using many methods such as reading the csv file, describing the data, and many more.
In this project I also work on data Visualization.